+++ CHANGELOG.md
@@ -4,8 +4,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2020-09-08
+
+### Add more styling options, alignment, length etc.
+
+- HeaderAlignment option
+- maxLength option
+
 ## [1.0.0] - 2020-08-30
 
 ### This is the first release of *\<Hierarchy Header\>*.
-
-*Short description of this release*